V11.5; Complete overhaul of timeMap() to make it cleaner.
In impulseResponse.py; Removed earlier timeMap() and timeMap_energy() in favour of just the one timeMap(), which can optionally return coeffs if kwargs passed.
In impulseResponse.py; _fs0() and _fs() added to replace impulseVec(). Computing only fs and not Fs (which is done as B@fs).
Energy and coeff computation for multiple forcings now expressed in terms of response to Fx, Fy, Fz by exploiting linearity.
To do: evolveField() to allow u(t) for a fixed t but set of modes